Build with -Wextra and -Werror=format-security
We'll need to suppress a few warnings, and fix some code that triggers
-Wsign-compare.

One -Wsign-compare is not fixed here because it's a real bug, to be
addressed separately.

We'll also manually add -Werror=format-security since this is amazingly
not included in -Wextra.

+++ lib/contrib/gvdb/gvdb-reader.c
@@ -418,7 +418,7 @@ gvdb_table_get_names (GvdbTable *table,
                 }
             }
 
-          else if (parent < n_names && names[parent] != NULL)
+          else if (parent < (guint32)n_names && names[parent] != NULL)
             {
               /* It's a non-root item whose parent was filled in already.
                *
